refactor: split eval property statement dispatch
