+++ crates/elephc-magician/src/interpreter/statements.rs
@@ -30,6 +30,7 @@ mod native_constructor_defaults;
 mod native_method_execution;
 mod native_static_dispatch;
 mod property_constant_validation;
+mod property_dispatch;
 mod property_validation;
 mod reference_writeback;
 mod reflection_instantiation;
@@ -65,6 +66,7 @@ pub(in crate::interpreter) use native_constructor_defaults::*;
 pub(in crate::interpreter) use native_method_execution::*;
 pub(in crate::interpreter) use native_static_dispatch::*;
 use property_constant_validation::*;
+use property_dispatch::*;
 pub(in crate::interpreter) use property_validation::*;
 pub(in crate::interpreter) use reference_writeback::*;
 pub(in crate::interpreter) use reflection_instantiation::*;
